+++ src/ts/components/playbackspeedselectbox.ts
@@ -20,11 +20,23 @@ export class PlaybackSpeedSelectBox extends SelectBox {
     this.addItem('1.5', '1.5x');
     this.addItem('2', '2x');
 
-    this.selectItem('1');
-
 
     this.onItemSelected.subscribe((sender: PlaybackSpeedSelectBox, value: string) => {
       player.setPlaybackSpeed(parseFloat(value));
+      this.selectItem(value);
     });
+
+    let setDefaultValue = (): void => {
+      let playbackSpeed = String(player.getPlaybackSpeed());
+      if (!this.selectItem(playbackSpeed)) {
+        playbackSpeed = '1';
+        this.selectItem(playbackSpeed);
+      }
+    };
+
+    setDefaultValue();
+
+    // when the player hits onReady again, adjust the playback speed selection with fallback to default 1
+    player.addEventHandler(player.EVENT.ON_READY, setDefaultValue);
   }
 }