+++ src/ts/components/subtitlesettings/subtitlesettingspanel.ts
@@ -11,9 +11,13 @@ import {CharacterEdgeSelectBox} from './characteredgeselectbox';
 import {SubtitleOverlay} from '../subtitleoverlay';
 import {Component, ComponentConfig} from '../component';
 import {UIInstanceManager} from '../../uimanager';
+import {SubtitleSettingsManager} from './subtitlesettingsmanager';
+import {SubtitleSettingsCloseButton} from './subtitlesettingsclosebutton';
+import {SubtitleSettingsResetButton} from './subtitlesettingsresetbutton';
 
 export interface SubtitleSettingsPanelConfig extends SettingsPanelConfig {
   overlay: SubtitleOverlay;
+  settingsPanel: SettingsPanel;
 }
 
 /**
@@ -28,18 +32,44 @@ export class SubtitleSettingsPanel extends SettingsPanel {
 
     this.overlay = config.overlay;
 
+    let manager = new SubtitleSettingsManager(this.overlay);
+
     this.config = this.mergeConfig(config, {
       components: <Component<ComponentConfig>[]>[
-        new SettingsPanelItem('Font family', new FontFamilySelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Font color', new FontColorSelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Font size', new FontSizeSelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Font opacity', new FontOpacitySelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Character edge', new CharacterEdgeSelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Background color', new BackgroundColorSelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Background opacity', new BackgroundOpacitySelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Window color', new WindowColorSelectBox({overlay: this.overlay})),
-        new SettingsPanelItem('Window opacity', new WindowOpacitySelectBox({overlay: this.overlay})),
-      ]}, this.config);
+        new SettingsPanelItem('Font size', new FontSizeSelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Font family', new FontFamilySelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Font color', new FontColorSelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Font opacity', new FontOpacitySelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Character edge', new CharacterEdgeSelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Background color', new BackgroundColorSelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Background opacity', new BackgroundOpacitySelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Window color', new WindowColorSelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem('Window opacity', new WindowOpacitySelectBox({
+          overlay: this.overlay, settingsManager: manager,
+        })),
+        new SettingsPanelItem(new SubtitleSettingsCloseButton({
+          subtitleSettingsPanel: this, settingsPanel: config.settingsPanel,
+        }), new SubtitleSettingsResetButton({
+          settingsManager: manager,
+        })),
+      ],
+    }, this.config);
   }
 
   configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {