+++ src/ts/main.ts
@@ -66,10 +66,10 @@ import {StringUtils} from './stringutils';
 import {PlayerUtils} from './playerutils';
 import {UIUtils} from './uiutils';
 import {BrowserUtils} from './browserutils';
-import {ColorUtils} from './colorutils';
 import {StorageUtils} from './storageutils';
 // HACK: gulp-tslint fails on unused files so we need this dummy import
 import {ArrayUtils as Dummy} from './utils'; // tslint:disable-line
+import {SubtitleSettingsResetButton} from './components/subtitlesettings/subtitlesettingsresetbutton';
 
 // Object.assign polyfill for ES5/IE9
 // https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Object/assign
@@ -107,7 +107,6 @@ if (typeof Object.assign !== 'function') {
   PlayerUtils,
   UIUtils,
   BrowserUtils,
-  ColorUtils,
   StorageUtils,
   // Components
   AdClickOverlay,
@@ -176,5 +175,6 @@ if (typeof Object.assign !== 'function') {
     SubtitleSettingsPanel,
     WindowColorSelectBox,
     WindowOpacitySelectBox,
+    SubtitleSettingsResetButton,
   },
 };