+++ CHANGELOG.md
@@ -9,11 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Adds a `VolumeToggleButton` to the small screen UI
 
+### Changed
+- Moved all subtitle styling to CSS (default subtitle style is not overwritten any longer)
+
 ### Fixed
 - Fix clearing of container components with `Container#removeComponents` (fixes sticky/duplicate subtitle issue)
 - Fix updating container components with `Container#updateComponents` (fixes empty subtitles in IE11)
 - Fix handling of duplicate subtitle cues (same text at same time) in `SubtitleOverlay` (fixes another sticky subtitle issue)
 - Fix clearing of recommendations in `RecommendationOverlay` (fixes duplicate recommendations issue)
+- Reset selected value in `ListSelector` when the items are cleared
+- Updating selected value in `PlaybackSpeedSelectBox` when player is ready
 - Fix video quality options for progressive streams (removed 'auto' option, preferred quality preselected)
 
 ## [2.7.1]